requirements.txt
discord.py==2.3.1
mvg==1.0.2
googlemaps==4.10.0
pytz=2023.3